moving to server_endpoints
