+++ src/cryptoadvance/specter/services/controller.py
@@ -3,10 +3,11 @@ from functools import wraps
 
 from flask import Blueprint
 from flask import current_app as app
-from flask import flash, redirect, render_template, request, url_for
+from flask import redirect, render_template, request, url_for
 from flask_babel import lazy_gettext as _
 from flask_login import current_user, login_required
 
+from ..server_endpoints import flash
 from ..services import ExtensionException
 
 logger = logging.getLogger(__name__)
