+++ src/cryptoadvance/specter/services/swan/controller.py
@@ -15,7 +15,7 @@ from flask.json import jsonify
 from flask_babel import lazy_gettext as _
 from flask_login import current_user, login_required
 
-from ...server import flash
+from ...server_endpoints import flash
 from ..controller import user_secret_decrypted_required
 from . import client as swan_client
 from .client import SwanApiException
