+++ src/cryptoadvance/specter/util/wallet_importer.py
@@ -1,17 +1,18 @@
 import json
 import logging
+
 import requests
-from cryptoadvance.specter.managers.wallet_manager import WalletManager
-
-from cryptoadvance.specter.specter_error import SpecterError
-
 from embit.descriptor import Descriptor
 from embit.descriptor import Key as DescriptorKey
 from embit.descriptor.arguments import AllowedDerivation
 from embit.liquid.descriptor import LDescriptor
-from cryptoadvance.specter.key import Key
 from flask_babel import lazy_gettext as _
+
+from ..key import Key
+from ..managers.wallet_manager import WalletManager
 from ..server import flash
+from ..server_endpoints import flash
+from ..specter_error import SpecterError
 
 logger = logging.getLogger(__name__)
